requirements.txt
fastapi
imdbpy
uvicorn
Levenshtein